# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AVANGENIO-TEST-API-DOCUMENTATION",
        "version": "3.0.0"
    },
    "paths": {
        "/api/game/CrearNuevoJuego": {
            "post": {
                "summary": "Endpoint para crear un nuevo juego",
                "operationId": "6a4dd7f15b8133a3dbe0a8bc18c2e21c",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "edad",
                        "in": "query",
                        "description": "Edad del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Juego creado correctamente"
                    },
                    "500": {
                        "description": "Error en la creacion del juego"
                    },
                    "400": {
                        "description": "Error en la validacion de los datos"
                    }
                }
            }
        },
        "/api/game/proponerCombinacion/{numero}": {
            "get": {
                "summary": "Endpoint para validar el numero propuesto",
                "operationId": "902749aca8c984d22cc9de319cc430b7",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "Nombre del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "El valor no es numerico"
                    },
                    "502": {
                        "description": "El valor debe ser de cuatro digitos"
                    },
                    "201": {
                        "description": "Este juego ya ha sido jugado con anterioridad"
                    },
                    "202": {
                        "description": "Game Over. Tiempo expirado"
                    },
                    "203": {
                        "description": "Game Win. Numero adivinado"
                    }
                }
            }
        },
        "/api/game/EliminarJuego/{id}": {
            "delete": {
                "summary": "Endpoint para eliminar un juego existente",
                "operationId": "96a6188ed5fce23946563b656428a380",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del juego a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Juego eliminado correctamente"
                    },
                    "500": {
                        "description": "El juego no existe."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}